sc_consensus_manual_seal

Module consensus

Extensions for manual seal to produce blocks valid for any runtime.

Modules

aura

Aura consensus data provider, This allows manual seal author blocks that are valid for runtimes that expect the aura-specific digests.

babe

BABE consensus data provider, This allows manual seal author blocks that are valid for runtimes that expect babe-specific digests.

timestamp

Mocked timestamp inherent, allows for manual seal to create blocks for runtimes that expect this inherent.

Traits

ConsensusDataProvider

Consensus data provider, manual seal uses this trait object for authoring blocks valid for any runtime.