sc_consensus_manual_seal

Function run_delayed_finalize

pub async fn run_delayed_finalize<B, CB, C, S>(
    __arg0: DelayedFinalizeParams<C, S>,
)
where
    B: BlockT + 'static,
    CB: ClientBackend<B> + 'static,
    C: HeaderBackend<B> + Finalizer<B, CB> + ProvideRuntimeApi<B> + BlockchainEvents<B> + 'static,
    S: SpawnNamed,

Creates a future for delayed finalization of manual sealed blocks.

The future needs to be spawned in the background alongside the run_manual_seal/run_instant_seal future. It is required that EngineCommand::SealNewBlock is send with finalize = false to not finalize blocks directly after building them. This also means that delayed finality can not be used with run_instant_seal_and_finalize.