Function sc_consensus_manual_seal::run_instant_seal

pub async fn run_instant_seal<B, CB, E, A, C, T>(
    block_import: BoxBlockImport<B, T>, 
    env: E, 
    backend: Arc<CB>, 
    pool: Arc<Pool<A>>, 
    select_chain: C, 
    inherent_data_providers: InherentDataProviders
) where
    A: ChainApi<Block = B, Hash = <B as BlockT>::Hash> + 'static,
    B: BlockT + 'static,
    CB: ClientBackend<B> + 'static,
    E: Environment<B> + 'static,
    E::Error: Display,
    <E::Proposer as Proposer<B>>::Error: Display,
    C: SelectChain<B> + 'static, 

runs the background authorship task for the instant seal engine. instant-seal creates a new block for every transaction imported into the transaction pool.