Function sc_consensus_manual_seal::run_instant_seal_and_finalize

pub async fn run_instant_seal_and_finalize<B, BI, CB, E, C, TP, SC, CIDP, P>(
    __arg0: InstantSealParams<B, BI, E, C, TP, SC, CIDP, P>
)where
    B: BlockT + 'static,
    BI: BlockImport<B, Error = Error, Transaction = TransactionFor<C, B>> + Send + Sync + 'static,
    C: HeaderBackend<B> + Finalizer<B, CB> + ProvideRuntimeApi<B> + 'static,
    CB: ClientBackend<B> + 'static,
    E: Environment<B> + 'static,
    E::Proposer: Proposer<B, Proof = P, Transaction = TransactionFor<C, B>>,
    SC: SelectChain<B> + 'static,
    TransactionFor<C, B>: 'static,
    TP: TransactionPool<Block = B>,
    CIDP: CreateInherentDataProviders<B, ()>,
    P: Send + Sync + 'static,

Runs the background authorship task for the instant seal engine. instant-seal creates a new block for every transaction imported into the transaction pool.

This function will finalize the block immediately as well. If you don’t want this behavior use run_instant_seal instead.