Module sc_consensus_manual_seal::rpc

RPC interface for the ManualSeal Engine.

Modules

gen_client
The generated client module.

Structs

CreatedBlock
return type of engine_createBlock

ManualSeal
A struct that implements the ManualSealApi.

ManualSealClient
The Client.

Enums

EngineCommand
Message sent to the background authorship task, usually by RPC.

Traits

ManualSealApi
RPC trait that provides methods for interacting with the manual-seal authorship task over rpc.

Functions

send_result
report any errors or successes encountered by the authorship task back to the rpc

Type Definitions

Sender
sender passed to the authorship task to report errors or successes.