Trait sc_consensus_manual_seal::rpc::ManualSealApiServer

pub trait ManualSealApiServer<Hash>: Sized + Send + Sync + 'static {
    fn create_block<'life0, 'async_trait>(
        &'life0 self, 
        create_empty: bool, 
        finalize: bool, 
        parent_hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<CreatedBlock<Hash>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn finalize_block<'life0, 'async_trait>(
        &'life0 self, 
        hash: Hash, 
        justification: Option<EncodedJustification>
    ) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn into_rpc(self) -> RpcModule<Self>
    where
        Hash: Send + Sync + 'static + DeserializeOwned + Serialize,
    { ... }
}

Server trait implementation for the ManualSealApi RPC API.

Required methods

fn create_block<'life0, 'async_trait>(
&'life0 self,
create_empty: bool,
finalize: bool,
parent_hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<CreatedBlock<Hash>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Instructs the manual-seal authorship task to create a new block

fn finalize_block<'life0, 'async_trait>(
&'life0 self,
hash: Hash,
justification: Option<EncodedJustification>
) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Instructs the manual-seal authorship task to finalize a block

Provided methods

fn into_rpc(self) -> RpcModule<Self> where
Hash: Send + Sync + 'static + DeserializeOwned + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.