Crate sc_consensus_manual_seal

A manual sealing engine: the engine listens for rpc calls to seal blocks and create forks. This is suitable for a testing environment.

Re-exports

pub use self::consensus::ConsensusDataProvider;

pub use self::rpc::CreatedBlock;

pub use self::rpc::EngineCommand;

Modules

consensus

Extensions for manual seal to produce blocks valid for any runtime.

rpc

RPC interface for the ManualSeal Engine.

Structs

FinalizeBlockParams

params for block finalization.

InstantSealParams

Params required to start the manual sealing authorship task.

ManualSealParams

Params required to start the instant sealing authorship task.

SealBlockParams

params for sealing a new block

Enums

Error

errors encountered by background block authorship task

Constants

MANUAL_SEAL_ENGINE_ID

The ConsensusEngineId of Manual Seal.

MAX_PROPOSAL_DURATION

max duration for creating a proposal in secs

Functions

finalize_block

finalizes a block in the backend with the given params.

import_queue

Instantiate the import queue for the manual seal consensus engine.

run_instant_seal

runs the background authorship task for the instant seal engine. instant-seal creates a new block for every transaction imported into the transaction pool.

run_manual_seal

Creates the background authorship task for the manual seal engine.

seal_block

seals a new block with the given params