Function sc_consensus_manual_seal::run_instant_seal

pub async fn run_instant_seal<B, BI, CB, E, C, A, SC, CIDP>(
    __arg0: InstantSealParams<B, BI, E, C, A, SC, CIDP>
) where
    A: ChainApi<Block = B> + 'static,
    B: BlockT + 'static,
    BI: BlockImport<B, Error = Error, Transaction = TransactionFor<C, B>> + Send + Sync + 'static,
    C: HeaderBackend<B> + Finalizer<B, CB> + ProvideRuntimeApi<B> + 'static,
    CB: ClientBackend<B> + 'static,
    E: Environment<B> + 'static,
    E::Proposer: Proposer<B, Transaction = TransactionFor<C, B>>,
    SC: SelectChain<B> + 'static,
    TransactionFor<C, B>: 'static,
    CIDP: CreateInherentDataProviders<B, ()>, 

runs the background authorship task for the instant seal engine. instant-seal creates a new block for every transaction imported into the transaction pool.