Module sc_consensus_manual_seal::rpc

RPC interface for the ManualSeal Engine.

Structs

CreatedBlock

return type of engine_createBlock

ManualSeal

A struct that implements the ManualSealApiServer.

Enums

EngineCommand

Message sent to the background authorship task, usually by RPC.

Traits

ManualSealApiClient

Client implementation for the ManualSealApi RPC API.

ManualSealApiServer

Server trait implementation for the ManualSealApi RPC API.

Functions

send_result

report any errors or successes encountered by the authorship task back to the rpc

Type Definitions

Sender

Sender passed to the authorship task to report errors or successes.